+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - `data` prop change will force update items only if `renderItem` is also updated
   - https://github.com/Shopify/flash-list/pull/453
+- Added render target info to `renderItem` callback
+  - https://github.com/Shopify/flash-list/pull/454
 
 ## [1.0.0] - 2022-06-17